start splitting main into modules
